Move the impls for UnresolvedTypeCheck to match the location of the other impls for each type.
