+++ sway-core/src/type_engine/type_id.rs
@@ -34,6 +34,24 @@ impl From<usize> for TypeId {
     }
 }
 
+impl UnresolvedTypeCheck for TypeId {
+    fn check_for_unresolved_types(&self) -> Vec<CompileError> {
+        use TypeInfo::*;
+        let span_override = if let TypeInfo::Ref(_, span) = look_up_type_id_raw(*self) {
+            Some(span)
+        } else {
+            None
+        };
+        match look_up_type_id(*self) {
+            UnknownGeneric { name } => vec![CompileError::UnableToInferGeneric {
+                ty: name.as_str().to_string(),
+                span: span_override.unwrap_or_else(|| name.span().clone()),
+            }],
+            _ => vec![],
+        }
+    }
+}
+
 impl TypeId {
     pub(crate) fn update_type(&mut self, type_mapping: &TypeMapping, span: &Span) {
         *self = match look_up_type_id(*self).matches_type_parameter(type_mapping) {
